+++ platformio/__main__.py
@@ -41,7 +41,7 @@ class PlatformioCLI(click.MultiCommand):  # pylint: disable=R0904
     def _handle_obsolate_command(name):
         if name in ("install", "list", "search", "show", "uninstall"):
             click.secho(
-                "Warning! `platformio %s` command is obsoleted and will be "
+                "Warning! `platformio %s` command is deprecated and will be "
                 "removed in the next release! Please use "
                 "`platformio platforms %s` instead." % (name, name),
                 fg="yellow"