+++ platformio/commands/run.py
@@ -74,16 +74,17 @@ def cli(ctx, environment, target, upload_port,  # pylint: disable=R0913,R0914
 class EnvironmentProcessor(object):
 
     RENAMED_OPTIONS = {
-        "LDF_CYCLIC": "LIB_DFCYCLIC",
+        "INSTALL_LIBS": "LIB_INSTALL",
         "IGNORE_LIBS": "LIB_IGNORE",
-        "USE_LIBS": "LIB_USE"
+        "USE_LIBS": "LIB_USE",
+        "LDF_CYCLIC": "LIB_DFCYCLIC"
     }
 
     def __init__(self, cmd_ctx, name, options,  # pylint: disable=R0913
                  targets, upload_port, verbose):
         self.cmd_ctx = cmd_ctx
         self.name = name
-        self.options = options
+        self.options = self._validate_options(options)
         self.targets = targets
         self.upload_port = upload_port
         self.verbose_level = int(verbose)
@@ -114,24 +115,29 @@ class EnvironmentProcessor(object):
 
         return not is_error
 
+    def _validate_options(self, options):
+        result = {}
+        for k, v in options.items():
+            _k = k.upper()
+            # process obsolete options
+            if _k in self.RENAMED_OPTIONS:
+                click.secho(
+                    "Warning! `%s` option is deprecated and will be "
+                    "removed in the next release! Please use "
+                    "`%s` instead." % (
+                        k, self.RENAMED_OPTIONS[_k].lower()),
+                    fg="yellow"
+                )
+                k = self.RENAMED_OPTIONS[_k].lower()
+            result[k] = v
+        return result
+
     def _get_build_variables(self):
         variables = ["PIOENV=" + self.name]
         if self.upload_port:
             variables.append("UPLOAD_PORT=%s" % self.upload_port)
         for k, v in self.options.items():
             k = k.upper()
-
-            # process obsolete options
-            if k in self.RENAMED_OPTIONS:
-                click.secho(
-                    "Warning! `%s` option is obsoleted and will be "
-                    "removed in the next release! Please use "
-                    "`%s` instead." % (
-                        k.lower(), self.RENAMED_OPTIONS[k].lower()),
-                    fg="yellow"
-                )
-                k = self.RENAMED_OPTIONS[k]
-
             if k == "TARGETS" or (k == "UPLOAD_PORT" and self.upload_port):
                 continue
             variables.append("%s=%s" % (k, v))
@@ -157,8 +163,8 @@ class EnvironmentProcessor(object):
 
         # install platform and libs dependencies
         _autoinstall_platform(self.cmd_ctx, platform)
-        if "install_libs" in self.options:
-            _autoinstall_libs(self.cmd_ctx, self.options['install_libs'])
+        if "lib_install" in self.options:
+            _autoinstall_libs(self.cmd_ctx, self.options['lib_install'])
 
         p = PlatformFactory.newPlatform(platform)
         return p.run(build_vars, build_targets, self.verbose_level)